+++ src/pages/Processos/ConvocacaoCandidatos/components/style.ts
@@ -20,16 +20,20 @@ export const StyledTable = styled(Table)`
 
   .ant-table-pagination {
     display: flex;
-    justify-content: flex-end; 
+    justify-content: flex-end;
+    .ant-pagination-item{
+    border-radius:5px
+    } 
   }
 
-  .ant-pagination-total-text {
-    margin-right: auto; 
-  }
   
+  
+   .ant-pagination-total-text {
+    margin-right: auto; 
+    color: #727679;
+  }
   box-shadow: 0px 6px 18px 0px rgba(0, 0, 0, 0.06);
 ` as typeof Table;
 
-
-
+
